requirements.txt
beautifulsoup4==4.12.3
blinker==1.7.0
cassandra-driver==3.29.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
geomet==0.2.1.post1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.2.0
MarkupSafe==2.1.5
Pebble==5.0.7
requests==2.31.0
six==1.16.0
soupsieve==2.5
urllib3==2.2.1
Werkzeug==3.0.1
